mybatis 데이터 연동 및 테스트
