+++ WebContent/test.jsp
@@ -4,27 +4,24 @@
 <html>
 <head>
 <meta charset="UTF-8">
+<script src='./lib/jquery-3.6.0.min.js'></script>
 <title>Insert title here</title>
-<script src='/Web_Semi/lib/jquery-3.6.0.min.js'></script>
-<script src='/Web_Semi/PJH/js/review.js'></script>
 </head>
 <body>
 <%
-	//기본 include 페이지
-	//String border = "./member/search.jsp";
-	String border = "./review.do";
-	//읽어올 페이지의 경로를 변수에 담아 두었다.
+
+	String review = "./review.do";
 	
-	request.setAttribute("border", border);
+	request.setAttribute("review", review);
 %>
-<div id='middle'>
-	<div id='border'></div>
+<div id="middle">
+		<div id="middle_main">
+		</div>
 </div>
 
+<script>
+	$('#middle_main').load('${review}', 'job=search');
+	//border의 영역에 servlet으로 불러온 결과를 load한다.
+</script>
 </body>
-
-	<script>
-		$('#border').load('${border}', 'job=search');
-		//border의 영역에 servlet으로 불러온 결과를 load한다.
-	</script>
 </html>
