+++ src/pjh/review.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+	PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+	
+<mapper namespace="review">
+
+	<select id="totList" parameterType="String" resultType="Integer">
+		select count(serial) totList from board 
+		where 	subject 	like '%${_parameter}%'
+		or 		doc		 	like '%${_parameter}%'
+	</select>
+	
+	<select id="search" parameterType="Integer" resultType="Integer">
+		select * from review
+	</select>
+
+</mapper>
